graduate work part1

+++ MyBookShopApp/src/main/java/com/example/MyBookShopApp/security/AddMoneyDto.java
@@ -2,13 +2,13 @@ package com.example.MyBookShopApp.security;
 
 public class AddMoneyDto
 {
-    private Integer sum;
+    private Double sum;
 
-    public Integer getSum() {
+    public Double getSum() {
         return sum;
     }
 
-    public void setSum(Integer sum) {
+    public void setSum(Double sum) {
         this.sum = sum;
     }
